🚚 Move tests to appropriate directory
