+++ tests/Documents/MarkdownDocument/MarkdownDocumentTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Aheenam\Mozhi\Test\Documents;
+namespace Aheenam\Mozhi\Test\Documents\MarkdownDocument;
 
 use Aheenam\Mozhi\Test\TestCase;
 use Illuminate\Support\Facades\Config;
@@ -12,7 +12,7 @@ class MarkdownDocumentTest extends TestCase
     public function it_creates_document_from_content()
     {
         Config::set('mozhi.default_template', 'default');
-        $content = file_get_contents(__DIR__.'/../tmp/contents/index.md');
+        $content = file_get_contents(__DIR__ . '/../../tmp/contents/index.md');
         $document = MarkdownDocument::fromContent($content);
 
         $this->assertInstanceOf(MarkdownDocument::class, $document);
@@ -23,7 +23,7 @@ class MarkdownDocumentTest extends TestCase
     /** @test */
     public function it_uses_the_correct_template_name()
     {
-        $content = file_get_contents(__DIR__.'/../tmp/contents/no-view/no-view.md');
+        $content = file_get_contents(__DIR__ . '/../../tmp/contents/no-view/no-view.md');
         $document = MarkdownDocument::fromContent($content);
 
         $this->assertEquals('nothing', $document->getTemplateName());
